Remove unused items and consolidate data fetching logic
