--- app/interactive/observationsviewer/PS_times.json
@@ -1,79 +1,0 @@
-[
-    {
-        "t": [
-            "01/11/2026 - 04PM PST",
-            "01/11/2026 - 05PM PST",
-            "01/11/2026 - 06PM PST",
-            "01/11/2026 - 07PM PST",
-            "01/11/2026 - 08PM PST",
-            "01/11/2026 - 09PM PST",
-            "01/11/2026 - 10PM PST",
-            "01/11/2026 - 11PM PST",
-            "01/12/2026 - 12AM PST",
-            "01/12/2026 - 01AM PST",
-            "01/12/2026 - 02AM PST",
-            "01/12/2026 - 03AM PST",
-            "01/12/2026 - 04AM PST",
-            "01/12/2026 - 05AM PST",
-            "01/12/2026 - 06AM PST",
-            "01/12/2026 - 07AM PST",
-            "01/12/2026 - 08AM PST",
-            "01/12/2026 - 09AM PST",
-            "01/12/2026 - 10AM PST",
-            "01/12/2026 - 11AM PST",
-            "01/12/2026 - 12PM PST",
-            "01/12/2026 - 01PM PST",
-            "01/12/2026 - 02PM PST",
-            "01/12/2026 - 03PM PST",
-            "01/12/2026 - 04PM PST",
-            "01/12/2026 - 05PM PST",
-            "01/12/2026 - 06PM PST",
-            "01/12/2026 - 07PM PST",
-            "01/12/2026 - 08PM PST",
-            "01/12/2026 - 09PM PST",
-            "01/12/2026 - 10PM PST",
-            "01/12/2026 - 11PM PST",
-            "01/13/2026 - 12AM PST",
-            "01/13/2026 - 01AM PST",
-            "01/13/2026 - 02AM PST",
-            "01/13/2026 - 03AM PST",
-            "01/13/2026 - 04AM PST",
-            "01/13/2026 - 05AM PST",
-            "01/13/2026 - 06AM PST",
-            "01/13/2026 - 07AM PST",
-            "01/13/2026 - 08AM PST",
-            "01/13/2026 - 09AM PST",
-            "01/13/2026 - 10AM PST",
-            "01/13/2026 - 11AM PST",
-            "01/13/2026 - 12PM PST",
-            "01/13/2026 - 01PM PST",
-            "01/13/2026 - 02PM PST",
-            "01/13/2026 - 03PM PST",
-            "01/13/2026 - 04PM PST",
-            "01/13/2026 - 05PM PST",
-            "01/13/2026 - 06PM PST",
-            "01/13/2026 - 07PM PST",
-            "01/13/2026 - 08PM PST",
-            "01/13/2026 - 09PM PST",
-            "01/13/2026 - 10PM PST",
-            "01/13/2026 - 11PM PST",
-            "01/14/2026 - 12AM PST",
-            "01/14/2026 - 01AM PST",
-            "01/14/2026 - 02AM PST",
-            "01/14/2026 - 03AM PST",
-            "01/14/2026 - 04AM PST",
-            "01/14/2026 - 05AM PST",
-            "01/14/2026 - 06AM PST",
-            "01/14/2026 - 07AM PST",
-            "01/14/2026 - 08AM PST",
-            "01/14/2026 - 09AM PST",
-            "01/14/2026 - 10AM PST",
-            "01/14/2026 - 11AM PST",
-            "01/14/2026 - 12PM PST",
-            "01/14/2026 - 01PM PST",
-            "01/14/2026 - 02PM PST",
-            "01/14/2026 - 03PM PST",
-            "01/14/2026 - 04PM PST"
-        ]
-    }
-]
